backend/requirements.txt
fastapi
pydantic[email]
aiofiles
sse-starlette
jinja2
jinja-markdown
requests
pyyaml
python-multipart
uvicorn[standard]
PyJWT
bcrypt
toml-config
pydbantic[sqlite]==0.0.47
loguru
